Backend/requirements.txt
fastapi
sqlmodel
uvicorn
faker
bcrypt
qrcode
itsdangerous
passlib[bcrypt]
pydantic[email]
python-multipart
python-jose[cryptography]
python-dotenv
pytest
httpx
